requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.4.0
async-timeout==4.0.3
asyncpg==0.31.0
attrs==25.4.0
Babel==2.9.1
backports-datetime-fromisoformat==2.0.3
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
environs==8.0.0
Flask==2.3.2
frozenlist==1.8.0
geographiclib==2.1
geopy==2.4.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
magic-filter==1.0.12
MarkupSafe==3.0.3
marshmallow==4.2.0
multidict==6.7.1
postgis==1.0.4
propcache==0.4.1
python-dotenv==1.2.1
pytz==2025.2
typing_extensions==4.15.0
Werkzeug==3.1.6
yarl==1.22.0
